requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.1
pytz==2024.1
